iterate on types
