+++ types/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./connectors/api";
 export * from "./connectors/api_handlers/create_connector";
 export * from "./core/data_source";
 export * from "./front/api_handlers/internal/agent_configuration";
